+++ servicecatalog-20210901/alibabacloud_servicecatalog20210901/models/_list_tag_options_request.py
@@ -9,6 +9,8 @@ class ListTagOptionsRequest(DaraModel):
     def __init__(
         self,
         filters: main_models.ListTagOptionsRequestFilters = None,
+        max_results: int = None,
+        next_token: str = None,
         page_number: int = None,
         page_size: int = None,
         sort_by: str = None,
@@ -16,6 +18,8 @@ class ListTagOptionsRequest(DaraModel):
     ):
         # The filter condition.
         self.filters = filters
+        self.max_results = max_results
+        self.next_token = next_token
         # The number of the page to return.
         # 
         # Pages start from page 1. Default value: 1.
@@ -46,6 +50,12 @@ class ListTagOptionsRequest(DaraModel):
         if self.filters is not None:
             result['Filters'] = self.filters.to_map()
 
+        if self.max_results is not None:
+            result['MaxResults'] = self.max_results
+
+        if self.next_token is not None:
+            result['NextToken'] = self.next_token
+
         if self.page_number is not None:
             result['PageNumber'] = self.page_number
 
@@ -65,6 +75,12 @@ class ListTagOptionsRequest(DaraModel):
         if m.get('Filters') is not None:
             temp_model = main_models.ListTagOptionsRequestFilters()
             self.filters = temp_model.from_map(m.get('Filters'))
+
+        if m.get('MaxResults') is not None:
+            self.max_results = m.get('MaxResults')
+
+        if m.get('NextToken') is not None:
+            self.next_token = m.get('NextToken')
 
         if m.get('PageNumber') is not None:
             self.page_number = m.get('PageNumber')